Moved Component and CompoentDeserializer to components directory
