+++ src/main/java/components/ComponentDeserializer.java
@@ -1,7 +1,8 @@
-package jade;
+package components;
 
 import com.google.gson.*;
 
+import components.Component;
 import java.lang.reflect.Type;
 
 public class ComponentDeserializer implements JsonSerializer<Component>, JsonDeserializer<Component> {
