+++ src/main/java/components/FontRenderer.java
@@ -1,7 +1,5 @@
 package components;
 
-import jade.Component;
-
 public class FontRenderer extends Component {
 
     @Override
