+++ src/main/java/components/Rigidbody.java
@@ -1,6 +1,5 @@
 package components;
 
-import jade.Component;
 import org.joml.Vector3f;
 import org.joml.Vector4f;
