+++ src/main/java/components/SpriteRenderer.java
@@ -1,7 +1,6 @@
 package components;
 
 import imgui.ImGui;
-import jade.Component;
 import jade.Transform;
 import org.joml.Vector2f;
 import org.joml.Vector4f;
