+++ src/main/java/jade/GameObject.java
@@ -1,6 +1,6 @@
 package jade;
 
-import org.joml.Vector2f;
+import components.Component;
 
 import java.util.ArrayList;
 import java.util.List;
