+++ src/main/java/jade/GameObjectDeserializer.java
@@ -2,6 +2,7 @@ package jade;
 
 import com.google.gson.*;
 
+import components.Component;
 import java.lang.reflect.Type;
 
 public class GameObjectDeserializer implements JsonDeserializer<GameObject> {
